Add GetCurrentHash tests
