+++ src/benchmarks/micro/libraries/System.IO.Hashing/Crc32_AppendPerf.cs
@@ -6,9 +6,9 @@ using BenchmarkDotNet.Attributes;
 
 namespace System.IO.Hashing.Tests
 {
-    public class Crc64_Perf : Crc_Perf<Crc64>
+    public class Crc32_AppendPerf : Crc_AppendPerf<Crc32>
     {
-        [Params(16, 256, 10240)]
+        [Params(16, 128, 10240)]
         public override int BufferSize { get; set; }
     }
 }
